feat: Complete Android build environment setup and APK compilation
- Updated React Native from 0.73.6 to 0.74.0 for compatibility
- Upgraded Android Gradle Plugin from 8.2.2 to 8.6.0
- Upgraded Gradle from 8.6 to 8.7
- Updated compileSdkVersion from 34 to 35
- Removed incompatible dependencies (react-native-reanimated, react-native-gesture-handler)
- Disabled Hermes JS engine to avoid heap space issues
- Increased Gradle heap size from 2GB to 4GB
- Created missing Android modules: InputSimulationModule, AccessibilityModule, SettingsModule, PermissionsModule
- Added QueuedAction data class for action queue management
- Created drawable resources for UI components and text selection
- Generated debug.keystore for APK signing
- Fixed Kotlin compilation errors and type mismatches
- Successfully compiled debug APK (247MB)
- Updated .gitignore to exclude build artifacts and Android SDK

+++ android/app/src/main/kotlin/com/fortniteassist/modules/AIInferenceModule.kt
@@ -111,7 +111,7 @@ class AIInferenceModule(reactContext: ReactApplicationContext) : ReactContextBas
                 putInt("inputWidth", modelInfo.inputWidth)
                 putInt("inputHeight", modelInfo.inputHeight)
                 putInt("numClasses", modelInfo.numClasses)
-                putDouble("confidenceThreshold", modelInfo.confidenceThreshold)
+                putDouble("confidenceThreshold", modelInfo.confidenceThreshold.toDouble())
                 putBoolean("supportsGpu", modelInfo.supportsGpu)
             }
